+++ controller/bindings/label-printer/label_printer.mli
@@ -1,8 +1,0 @@
-type label =
-  { machine_id: string
-  ; mac_1: string
-  ; mac_2: string
-  }
-[@@deriving sexp]
-
-val print : url : string -> label -> unit Lwt.t