+++ controller/server/view/label_page.mli
@@ -1,3 +1,0 @@
-val html :
-  Label_printer.label
-  -> [> Html_types.html ] Tyxml.Html.elt